+++ ContaSimples.java
@@ -1,5 +1,5 @@
 public class ContaSimples extends Conta {
-  public ContaSimples(String nome, String endereco, int cpf, int data, float saldo){
+  public ContaSimples(String nome, String endereco, int cpf, int data, double saldo){
     super(nome, endereco, cpf, data, saldo);
   }
 }
